bq-to-arrow-stream/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Jonas Eyob <jonas.eyob@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyarrow = "^12.0.0"
google-cloud-bigquery-storage = "^2.19.1"
pydantic = "^1.10.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
